sospharma-backend/requirements.txt
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.1
requests==2.31.0
